Move spec-oriented test to its own file.
Spec-oriented tests should follow the same convention.
