+++ ably/realtime_conn_test.go
@@ -2,14 +2,11 @@ package ably_test
 
 import (
 	"fmt"
-	"io"
-	"net/url"
 	"testing"
 	"time"
 
 	"github.com/ably/ably-go/ably"
 	"github.com/ably/ably-go/ably/ablytest"
-	"github.com/ably/ably-go/ably/internal/ablyutil"
 	"github.com/ably/ably-go/ably/proto"
 )
 
@@ -193,121 +190,6 @@ func TestRealtimeConn_ReceiveTimeout(t *testing.T) {
 
 	if expected, got := ably.StateConnDisconnected, state.State; expected != got {
 		t.Fatalf("expected %v, got %v", expected, got)
-	}
-}
-
-func TestRealtimeConn_RTN15a_ReconnectOnEOF(t *testing.T) {
-	t.Parallel()
-
-	doEOF := make(chan struct{}, 1)
-
-	app, client := ablytest.NewRealtimeClient(&ably.ClientOptions{
-		NoConnect: true,
-		Dial: func(protocol string, u *url.URL) (proto.Conn, error) {
-			c, err := ablyutil.DialWebsocket(protocol, u)
-			return protoConnWithFakeEOF{Conn: c, doEOF: doEOF}, err
-		},
-	})
-	defer safeclose(t, client, app)
-
-	if err := ablytest.Wait(client.Connection.Connect()); err != nil {
-		t.Fatalf("Connect=%s", err)
-	}
-
-	channel := client.Channels.Get("channel")
-
-	if err := ablytest.Wait(channel.Attach()); err != nil {
-		t.Fatal(err)
-	}
-
-	sub, err := channel.Subscribe()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	stateChanges := make(chan ably.State, 16)
-	client.Connection.On(stateChanges)
-
-	doEOF <- struct{}{}
-
-	var state ably.State
-
-	select {
-	case state = <-stateChanges:
-	case <-time.After(50 * time.Millisecond):
-		t.Fatal("didn't transition on EOF")
-	}
-
-	if expected, got := ably.StateConnDisconnected, state; expected != got.State {
-		t.Fatalf("expected transition to %v, got %v", expected, got)
-	}
-
-	// Publish a message to the channel through REST. If connection recovery
-	// succeeds, we should then receive it without reattaching.
-
-	rest, err := ably.NewRestClient(app.Options())
-	if err != nil {
-		t.Fatal(err)
-	}
-	err = rest.Channels.Get("channel", nil).Publish("name", "data")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	select {
-	case state = <-stateChanges:
-	case <-time.After(50 * time.Millisecond):
-		t.Fatal("didn't reconnect")
-	}
-
-	if expected, got := ably.StateConnConnecting, state; expected != got.State {
-		t.Fatalf("expected transition to %v, got %v", expected, got)
-	}
-
-	select {
-	case state = <-stateChanges:
-	case <-time.After(ablytest.Timeout):
-		t.Fatal("didn't transition from CONNECTING")
-	}
-
-	if expected, got := ably.StateConnConnected, state; expected != got.State {
-		t.Fatalf("expected transition to %v, got %v", expected, got)
-	}
-
-	select {
-	case msg := <-sub.MessageChannel():
-		if expected, got := "data", msg.Data; expected != got {
-			t.Fatalf("expected message with data %v, got %v", expected, got)
-		}
-	case <-time.After(ablytest.Timeout):
-		t.Fatal("expected message after connection recovery; got none")
-	}
-}
-
-type protoConnWithFakeEOF struct {
-	proto.Conn
-	doEOF <-chan struct{}
-}
-
-func (c protoConnWithFakeEOF) Receive(deadline time.Time) (*proto.ProtocolMessage, error) {
-	type result struct {
-		msg *proto.ProtocolMessage
-		err error
-	}
-
-	received := make(chan result, 1)
-
-	go func() {
-		msg, err := c.Conn.Receive(deadline)
-		received <- result{msg: msg, err: err}
-
-	}()
-
-	select {
-	case r := <-received:
-		return r.msg, r.err
-	case <-c.doEOF:
-		return nil, io.EOF
 	}
 }
